# Backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Multi-Tenant Security Management Platform API",
    "version": "1.0.0",
    "description": "Production-ready REST API for Multi-Tenant Security Platform.\nEnforces strict database-level multi-tenant isolation, Role-Based Access Control (RBAC), JWT token revocation, append-only audit logging, and input sanitization.\n\n### Base Configuration\n- **Server Port**: 5000\n- **Environment**: development\n- **Database**: Multi_Tenant\n\n### Roles & Access Matrix\n- **ADMIN**: Complete authority over tenant data, user management, campaigns, security events, and audit logs.\n- **MANAGER**: Full control over campaigns, security events, user assignments, and viewing users.\n- **USER**: Read-only view of assigned campaigns and security events. No access to audit logs or user management."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Environment Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token obtained from POST /api/auth/login"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Authenticate User & Receive JWT",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "admin@acme.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "AdminPassword123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated successfully"
          },
          "401": {
            "description": "Invalid email or password"
          }
        }
      }
    },
...
  }
}